Some Gadget support and various refactoring (#27)
* various gadget boilerplate

* move ConfigBorn and ConfigAttack to Shared folder

* Refactor ScriptEntity to IScriptEntity

* ConfigGadget part 1

* SerializationBinder bind types via reflection

* BaseBornType -> ConfigBornType, AttackInfo -> ConfigAttackInfo

* ConfigGadget part 2

* refactor various fields int -> uint

* implement GadgetAbilityManager

* better namespace reflection

* even better namespace reflection

* refactor Command layout

* add gadgetId to Gadget constructors

* small fixes

+++ src/GameServer/Packet/Recv/HandleSceneTransToPointReq.cs
@@ -34,7 +34,7 @@ namespace Weedwacker.GameServer.Packet.Recv
                             bp.tranRot.FirstOrDefault(x => x.Key.Contains('z')).Value);
                     }
                     await session.Player.World.TransferPlayerToSceneAsync(session.Player, EnterReason.TransPoint,
-                        req.SceneId == (uint)session.Player.SceneId ? EnterType.Goto : EnterType.Jump, (int)req.SceneId,
+                        req.SceneId == (uint)session.Player.SceneId ? EnterType.Goto : EnterType.Jump, req.SceneId,
                         pos, rot, false);
                 }
                 else
