+++ src/GameServer/Packet/Send/PacketAvatarSkillChangeNotify.cs
@@ -7,16 +7,16 @@ namespace Weedwacker.GameServer.Packet.Send
 {
     internal class PacketAvatarSkillChangeNotify : BasePacket
     {
-        public PacketAvatarSkillChangeNotify(Avatar avatar, int depotId, int skillId, int oldLevel, int curLevel) : base(OpCode.AvatarSkillChangeNotify)
+        public PacketAvatarSkillChangeNotify(Avatar avatar, uint depotId, uint skillId, uint oldLevel, uint curLevel) : base(OpCode.AvatarSkillChangeNotify)
         {
             AvatarSkillChangeNotify proto = new AvatarSkillChangeNotify()
             {
                 AvatarGuid = avatar.Guid,
-                EntityId = (uint)avatar.EntityId,
-                SkillDepotId = (uint)depotId,
-                AvatarSkillId = (uint)skillId,
-                OldLevel = (uint)oldLevel,
-                CurLevel = (uint)curLevel
+                EntityId = avatar.EntityId,
+                SkillDepotId = depotId,
+                AvatarSkillId = skillId,
+                OldLevel = oldLevel,
+                CurLevel = curLevel
             };
 
             Data = proto.ToByteArray();
