+++ src/GameServer/Packet/Send/PacketAvatarSkillInfoNotify.cs
@@ -5,14 +5,14 @@ namespace Weedwacker.GameServer.Packet.Send
 {
     internal class PacketAvatarSkillInfoNotify : BasePacket
     {
-        public PacketAvatarSkillInfoNotify(ulong avatarGuid, IDictionary<int, int> skillExtraChargeMap) : base(Enums.OpCode.AvatarSkillInfoNotify)
+        public PacketAvatarSkillInfoNotify(ulong avatarGuid, IDictionary<uint, uint> skillExtraChargeMap) : base(Enums.OpCode.AvatarSkillInfoNotify)
         {
             var proto = new AvatarSkillInfoNotify()
             {
                 Guid = avatarGuid
             };
 
-            skillExtraChargeMap.AsParallel().ForAll(w => proto.SkillMap.Add((uint)w.Key, new AvatarSkillInfo() { MaxChargeCount = (uint)w.Value }));
+            skillExtraChargeMap.AsParallel().ForAll(w => proto.SkillMap.Add(w.Key, new AvatarSkillInfo() { MaxChargeCount = w.Value }));
 
             Data = proto.ToByteArray();
         }
