+++ src/GameServer/Packet/Send/PacketDelTeamEntityNotify.cs
@@ -5,11 +5,11 @@ namespace Weedwacker.GameServer.Packet.Send
 {
     internal class PacketDelTeamEntityNotify : BasePacket
     {
-        public PacketDelTeamEntityNotify(int sceneId, uint teamEntityId) : base(Enums.OpCode.DelTeamEntityNotify)
+        public PacketDelTeamEntityNotify(uint sceneId, uint teamEntityId) : base(Enums.OpCode.DelTeamEntityNotify)
         {
             DelTeamEntityNotify proto = new DelTeamEntityNotify()
             {
-                SceneId = (uint)sceneId,
+                SceneId = sceneId,
             };
             proto.DelEntityIdList.Add(teamEntityId);
