+++ src/GameServer/Packet/Send/PacketOpenStateChangeNotify.cs
@@ -5,21 +5,21 @@ namespace Weedwacker.GameServer.Packet.Send
 {
     internal class PacketOpenStateChangeNotify : BasePacket
     {
-        public PacketOpenStateChangeNotify(int openState, int value) : base(Enums.OpCode.OpenStateChangeNotify)
+        public PacketOpenStateChangeNotify(uint openState, int value) : base(Enums.OpCode.OpenStateChangeNotify)
         {
             OpenStateChangeNotify proto = new OpenStateChangeNotify();
-            proto.OpenStateMap.Add((uint)openState, (uint)value);
+            proto.OpenStateMap.Add(openState, (uint)value);
 
             Data = proto.ToByteArray();
         }
 
-        public PacketOpenStateChangeNotify(IEnumerable<Tuple<int, int>> openStates) : base(Enums.OpCode.OpenStateChangeNotify)
+        public PacketOpenStateChangeNotify(IEnumerable<Tuple<uint, int>> openStates) : base(Enums.OpCode.OpenStateChangeNotify)
         {
 
             OpenStateChangeNotify proto = new OpenStateChangeNotify();
             foreach (var openState in openStates)
             {
-                proto.OpenStateMap.Add((uint)openState.Item1, (uint)openState.Item2);
+                proto.OpenStateMap.Add(openState.Item1, (uint)openState.Item2);
             }
 
             Data = proto.ToByteArray();
