+++ src/GameServer/Packet/Send/PacketPlayerEnterSceneNotify.cs
@@ -11,7 +11,7 @@ namespace Weedwacker.GameServer.Packet.Send
 {
     internal class PacketPlayerEnterSceneNotify : BasePacket
     {
-        public PacketPlayerEnterSceneNotify(Player target, EnterType type, EnterReason reason, Scene newScene, Vector3 newPos, int prevSceneId = default, Vector3 prevPos = default) : base(OpCode.PlayerEnterSceneNotify)
+        public PacketPlayerEnterSceneNotify(Player target, EnterType type, EnterReason reason, Scene newScene, Vector3 newPos, uint prevSceneId = default, Vector3 prevPos = default) : base(OpCode.PlayerEnterSceneNotify)
         {
             target.SceneLoadState = SceneLoadState.LOADING;
             Random rand = new();
@@ -19,11 +19,11 @@ namespace Weedwacker.GameServer.Packet.Send
 
             PlayerEnterSceneNotify proto = new PlayerEnterSceneNotify()
             {
-                SceneId = (uint)newScene.SceneId,
+                SceneId = newScene.SceneId,
                 Pos = new() { X = newPos.X, Y = newPos.Y, Z = newPos.Z },
                 SceneBeginTime = (ulong)DateTimeOffset.Now.ToUnixTimeMilliseconds(),
                 Type = type,
-                TargetUid = (uint)target.GameUid,
+                TargetUid = target.GameUid,
                 EnterSceneToken = target.EnterSceneToken,
                 WorldLevel = (uint)target.World.WorldLevel, // we need to use the world level of the world the player is in, NOT their own world level
                 EnterReason = (uint)reason,
@@ -31,11 +31,11 @@ namespace Weedwacker.GameServer.Packet.Send
                 SceneTransaction = newScene.SceneId + "-" + target.GameUid + "-" + (int)(DateTimeOffset.Now.ToUnixTimeMilliseconds() / 1000) + "-" + 18402
             };
 
-            if (newScene.DungeonData != null) proto.DungeonId = (uint)newScene.DungeonData.id;
+            if (newScene.DungeonData != null) proto.DungeonId = newScene.DungeonData.id;
             if (reason == EnterReason.Login) proto.IsFirstLoginEnterScene = true;
             if (prevSceneId != default)
             {
-                proto.PrevSceneId = (uint)prevSceneId;
+                proto.PrevSceneId = prevSceneId;
                 proto.PrevPos = new() { X = prevPos.X, Y = prevPos.Y, Z = prevPos.Z };
             }
             //TODO SkipUI and SceneTagList
