+++ src/GameServer/Systems/Social/Friendship.cs
@@ -6,22 +6,22 @@ namespace Weedwacker.GameServer.Systems.Social
 {
     internal class Friendship
     {
-        [BsonElement] public int FriendUid { get; private set; }
+        [BsonElement] public uint FriendUid { get; private set; }
         [BsonElement] public PlayerProfile FriendProfile { get; private set; }
         public FriendBrief ToProto()
         {
             FriendBrief proto = new FriendBrief()
             {
-                Uid = (uint)FriendUid,
+                Uid = FriendUid,
                 Nickname = FriendProfile.Nickname,
-                Level = (uint)FriendProfile.Level,
+                Level = FriendProfile.Level,
                 ProfilePicture = FriendProfile.HeadImage,
-                WorldLevel = (uint)FriendProfile.WorldLevel,
+                WorldLevel = FriendProfile.WorldLevel,
                 Signature = FriendProfile.Signature,
                 OnlineState = GameServer.OnlinePlayers.ContainsKey(FriendUid) ? FriendOnlineState.Online : FriendOnlineState.Disconnect,
                 LastActiveTime = (uint)FriendProfile.LastActiveTime,
-                NameCardId = (uint)FriendProfile.NameCardId,
-                Param = (uint)FriendProfile.DaysSinceLastLogin,
+                NameCardId = FriendProfile.NameCardId,
+                Param = FriendProfile.DaysSinceLastLogin,
                 IsGameSource = true,
                 PlatformType = PlatformType.Editor,
             };
